+++ backend/projektRoz/apiViews/__init__.py
@@ -1,0 +1,10 @@
+from .AddressApiView import AddressApiView
+from .ChildrenApiView import ChildrenApiView
+from .FatherApiView import FatherApiView
+from .FosterCarerApiView import FosterCarerApiView
+from .MotherApiView import MotherApiView
+from .NotesApiView import NotesApiView
+from .SiblingsApiView import SiblingsApiView
+from .CategoryApiView import CategoryApiView
+from .DocumentsApiView import DocumentsApiView
+from .AddressRegisteredApiView import AddressRegisteredApiView
